Changed double to long for move score in AlphaBeta.

+++ src/main/java/com/igorternyuk/engine/player/ai/BoardEvaluator.java
@@ -6,5 +6,5 @@ import com.igorternyuk.engine.board.Board;
  * Created by igor on 06.12.18.
  */
 public interface BoardEvaluator {
-    double evaluate(Board board, int depth);
+    long evaluate(Board board, int depth);
 }
